Dependency changes, URL fixes, `dccMarkdown` formatting.
This commit fixes the dataset links for `dash-bio` chapters to bring them in line with `master` in preparation for merging `dashr-squashed`. It also addresses version changes of packages on CRAN, and fixes dependencies so the docs can be deployed. Also fixes `dccMarkdown` unusual behaviour with formatting in the `dashCoreComponents` chapter.